refactor(auth-components): move auth auth components to features folder
