+++ client/src/components/index.js
@@ -1,12 +1,5 @@
 export { default as Navbar } from "./Navbar/Navbar";
 
-//Sign In Folder
-export { default as Content } from "./SignIn/Content";
-export { default as ForgotPassword } from "./SignIn/ForgotPassword";
-export { default as SignIn } from "./SignIn/SignIn";
-export { default as SignInCard } from "./SignIn/SignInCard";
-
-//Components
 export { default as ErrorMessage } from "./ErrorMessage/index";
 export { default as Form } from "./Form/Form";
 export { default as FormButton } from "./Form/FormButton";
