+++ client/src/features/auth/SignUp.jsx
@@ -1,0 +1,10 @@
+const SignUp = () => {
+  return (
+    <div>
+      <h1>Sign Up</h1>
+      {/* your signup form here */}
+    </div>
+  );
+};
+
+export default SignUp;
