+++ client/src/features/auth/index.js
@@ -1,0 +1,7 @@
+export { default as Content } from "features/auth/Content";
+export { default as ForgotPassword } from "features/auth/ForgotPassword";
+export { default as SignIn } from "features/auth/SignIn/SignIn";
+export { default as SignInCard } from "features/auth/SignIn/SignInCard";
+export { default as Card } from "features/auth/Card";
+export { default as SignUp } from "features/auth/SignUp";
+export { default as AuthCard } from "features/auth/AuthCard";
