+++ client/src/pages/index.js
@@ -1,4 +1,4 @@
-export { default as Home } from "./Home/index";
+export { default as Home } from "./Home";
 export { default as JobBoard } from "./JobBoard/index";
 export { default as Overview } from "./Overview/index";
 export { default as Networking } from "./Networking/index";
